refactor: extract compiler policies
